requirements.txt
colorama==0.4.4
googlesearch-python==1.1.0
requests==2.25.1



